requirements.txt
git+https://github.com/Rapptz/discord.py@rewrite
ruamel.yaml
uvloop
motor
aioredis